+++ packages/web/src/app/[domain]/browse/[...path]/page.tsx
@@ -58,7 +58,7 @@ export default async function BrowsePage({
     if (isServiceError(reposResponse)) {
         throw new ServiceErrorException(reposResponse);
     }
-    const repo = reposResponse.List.Repos.find(r => r.Repository.Name === repoName);
+    const repo = reposResponse.repos.find(r => r.name === repoName);
 
     if (pathType === 'tree') {
         // @todo : proper tree handling
@@ -82,7 +82,7 @@ export default async function BrowsePage({
                         <div className="bg-accent py-1 px-2 flex flex-row">
                             <FileHeader
                                 fileName={path}
-                                repo={repo.Repository}
+                                repo={repo}
                                 branchDisplayName={revisionName}
                             />
                         </div>
