+++ packages/web/src/app/[domain]/search/components/filterPanel/index.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { FileIcon } from "@/components/ui/fileIcon";
-import { Repository, SearchResultFile } from "@/lib/types";
+import { Repository, SearchResultFile } from "@/features/search/types";
 import { cn, getRepoCodeHostInfo } from "@/lib/utils";
 import { LaptopIcon } from "@radix-ui/react-icons";
 import Image from "next/image";
