+++ packages/web/src/features/search/listReposApi.ts
@@ -1,7 +1,7 @@
-import { listRepositoriesResponseSchema } from "../../lib/schemas";
-import { invalidZoektResponse, ServiceError, unexpectedError } from "../../lib/serviceError";
-import { ListRepositoriesResponse } from "../../lib/types";
+import { invalidZoektResponse, ServiceError } from "../../lib/serviceError";
+import { ListRepositoriesResponse } from "./types";
 import { zoektFetch } from "./zoektClient";
+import { zoektListRepositoriesResponseSchema } from "./zoektSchema";
 
 
 export const listRepositories = async (orgId: number): Promise<ListRepositoriesResponse | ServiceError> => {
@@ -29,11 +29,16 @@ export const listRepositories = async (orgId: number): Promise<ListRepositoriesR
     }
 
     const listBody = await listResponse.json();
-    const parsedListResponse = listRepositoriesResponseSchema.safeParse(listBody);
-    if (!parsedListResponse.success) {
-        console.error(`Failed to parse zoekt response. Error: ${parsedListResponse.error}`);
-        return unexpectedError(`Something went wrong while parsing the response from zoekt`);
-    }
 
-    return parsedListResponse.data;
+    const parser = zoektListRepositoriesResponseSchema.transform(({ List }) => ({
+        repos: List.Repos.map((repo) => ({
+            name: repo.Repository.Name,
+            url: repo.Repository.URL,
+            source: repo.Repository.Source,
+            branches: repo.Repository.Branches?.map((branch) => branch.Name) ?? [],
+            rawConfig: repo.Repository.RawConfig ?? undefined,
+        }))
+    } satisfies ListRepositoriesResponse));
+
+    return parser.parse(listBody);
 }
