+++ packages/web/src/features/search/types.ts
@@ -1,0 +1,25 @@
+import {
+    fileSourceResponseSchema,
+    listRepositoriesResponseSchema,
+    locationSchema,
+    searchRequestSchema,
+    searchResponseSchema,
+    rangeSchema,
+    fileSourceRequestSchema,
+    symbolSchema,
+} from "./schemas";
+import { z } from "zod";
+
+export type SearchRequest = z.infer<typeof searchRequestSchema>;
+export type SearchResponse = z.infer<typeof searchResponseSchema>;
+export type SearchResultRange = z.infer<typeof rangeSchema>;
+export type SearchResultLocation = z.infer<typeof locationSchema>;
+export type SearchResultFile = SearchResponse["files"][number];
+export type SearchResultChunk = SearchResultFile["chunks"][number];
+export type Symbol = z.infer<typeof symbolSchema>;
+
+export type ListRepositoriesResponse = z.infer<typeof listRepositoriesResponseSchema>;
+export type Repository = ListRepositoriesResponse["repos"][number];
+
+export type FileSourceRequest = z.infer<typeof fileSourceRequestSchema>;
+export type FileSourceResponse = z.infer<typeof fileSourceResponseSchema>;
