+++ packages/web/src/lib/extensions/searchResultHighlightExtension.ts
@@ -1,6 +1,6 @@
 import { EditorSelection, Extension, StateEffect, StateField, Text, Transaction } from "@codemirror/state";
 import { Decoration, DecorationSet, EditorView } from "@codemirror/view";
-import { SearchResultRange } from "../types";
+import { SearchResultRange } from "@/features/search/types";
 
 const setMatchState = StateEffect.define<{
     selectedMatchIndex: number,
