+++ packages/web/src/lib/types.ts
@@ -1,24 +1,8 @@
 import { z } from "zod";
-import { fileSourceRequestSchema, fileSourceResponseSchema, listRepositoriesResponseSchema, locationSchema, rangeSchema, repositorySchema, repositoryQuerySchema, searchRequestSchema, searchResponseSchema, symbolSchema, getVersionResponseSchema } from "./schemas";
+import { getVersionResponseSchema, repositoryQuerySchema } from "./schemas";
 import { tenancyModeSchema } from "@/env.mjs";
 
 export type KeymapType = "default" | "vim";
-
-export type SearchRequest = z.infer<typeof searchRequestSchema>;
-export type SearchResponse = z.infer<typeof searchResponseSchema>;
-
-export type SearchResultFile = SearchResponse["files"][number];
-export type SearchResultChunk = SearchResultFile["chunks"][number];
-export type SearchResultRange = z.infer<typeof rangeSchema>;
-export type SearchResultLocation = z.infer<typeof locationSchema>;
-
-export type FileSourceRequest = z.infer<typeof fileSourceRequestSchema>;
-export type FileSourceResponse = z.infer<typeof fileSourceResponseSchema>;
-
-export type ListRepositoriesResponse = z.infer<typeof listRepositoriesResponseSchema>;
-export type Repository = z.infer<typeof repositorySchema>;
-export type RepositoryQuery = z.infer<typeof repositoryQuerySchema>;
-export type Symbol = z.infer<typeof symbolSchema>;
 
 export type GetVersionResponse = z.infer<typeof getVersionResponseSchema>;
 
@@ -28,3 +12,4 @@ export enum SearchQueryParams {
 }
 
 export type TenancyMode = z.infer<typeof tenancyModeSchema>;
+export type RepositoryQuery = z.infer<typeof repositoryQuerySchema>;
